Move deprecated code into their own folders
