+++ src/services/deprecated/external-api-helper.ts
@@ -1,0 +1,83 @@
+import { Movie, SearchRequest, TVShow } from '@universalmediaserver/node-imdb-api';
+import * as _ from 'lodash';
+import * as episodeParser from 'episode-parser';
+import * as natural from 'natural';
+
+import { IMDbIDNotFoundError } from '../../helpers/customErrors';
+import { MediaMetadataInterface } from '../../models/MediaMetadata';
+import omdbAPI from '../omdb-api';
+import { mapper } from '../../utils/data-mapper';
+
+/**
+ * @deprecated see getFromOMDbAPIV2
+ */
+export const getFromOMDbAPI = async(imdbId?: string, searchRequest?: SearchRequest): Promise<MediaMetadataInterface> => {
+  if (!imdbId && !searchRequest) {
+    throw new Error('All parameters were falsy');
+  }
+
+  /**
+   * We need the IMDb ID for the OMDb API get request below so here we get it.
+   * Along the way, if the result is an episode, we also instruct our episode
+   * processor to asynchronously add the other episodes for that series to the
+   * queue.
+   */
+  if (!imdbId) {
+    const parsedFilename = episodeParser(searchRequest.name);
+    const isTVEpisode = Boolean(parsedFilename && parsedFilename.show && parsedFilename.season && parsedFilename.episode);
+    if (isTVEpisode) {
+      searchRequest.name = parsedFilename.show;
+      searchRequest.reqtype = 'series';
+      const tvSeriesInfo = await omdbAPI.get(searchRequest);
+
+      if (tvSeriesInfo && tvSeriesInfo instanceof TVShow) {
+        const allEpisodes = await tvSeriesInfo.episodes();
+        const currentEpisode = _.find(allEpisodes, { season: parsedFilename.season, episode: parsedFilename.episode });
+        if (!currentEpisode) {
+          throw new IMDbIDNotFoundError();
+        }
+
+        imdbId = currentEpisode.imdbid;
+      }
+    }
+
+    if (!imdbId) {
+      searchRequest.reqtype = 'movie';
+      let searchResults;
+      try {
+        searchResults = await omdbAPI.search(searchRequest);
+      } catch (e) {
+        console.error(e);
+        return null;
+      }
+      // find the best search results utilising the Jaro-Winkler distance metric
+      const searchResultStringDistance = searchResults.results.map(result => natural.JaroWinklerDistance(searchRequest.name, result.title));
+      const bestSearchResultKey = _.indexOf(searchResultStringDistance, _.max(searchResultStringDistance));
+
+      const searchResult = searchResults.results[bestSearchResultKey] as Movie;
+      if (!searchResult) {
+        throw new IMDbIDNotFoundError();
+      }
+
+      imdbId = searchResult.imdbid;
+    }
+  }
+
+  const imdbData = await omdbAPI.get({ id: imdbId });
+  if (!imdbData) {
+    return null;
+  }
+
+  let metadata;
+  if (imdbData.type === 'movie') {
+    metadata = mapper.parseOMDbAPIMovieResponse(imdbData);
+  } else if (imdbData.type === 'series') {
+    metadata = mapper.parseOMDbAPISeriesResponse(imdbData);
+  } else if (imdbData.type === 'episode') {
+    metadata = mapper.parseOMDbAPIEpisodeResponse(imdbData);
+  } else {
+    throw new Error('Received a type we did not expect');
+  }
+
+  return metadata;
+};
