+++ src/services/external-api-helper.ts
@@ -7,7 +7,7 @@ import * as natural from 'natural';
 import imdbAPI from '../services/omdb-api';
 import osAPI from '../services/opensubtitles';
 
-import { IMDbIDNotFoundError, ValidationError } from '../helpers/customErrors';
+import { ValidationError } from '../helpers/customErrors';
 import FailedLookups, { FailedLookupsInterface } from '../models/FailedLookups';
 import { MediaMetadataInterface } from '../models/MediaMetadata';
 import SeriesMetadata, { SeriesMetadataInterface } from '../models/SeriesMetadata';
@@ -35,80 +35,6 @@ export interface OpenSubtitlesValidation {
 interface SortByFilter {
   startYear: number;
 }
-
-/**
- * @deprecated see getFromOMDbAPIV2
- */
-export const getFromOMDbAPI = async(imdbId?: string, searchRequest?: SearchRequest): Promise<MediaMetadataInterface> => {
-  if (!imdbId && !searchRequest) {
-    throw new Error('All parameters were falsy');
-  }
-
-  /**
-   * We need the IMDb ID for the OMDb API get request below so here we get it.
-   * Along the way, if the result is an episode, we also instruct our episode
-   * processor to asynchronously add the other episodes for that series to the
-   * queue.
-   */
-  if (!imdbId) {
-    const parsedFilename = episodeParser(searchRequest.name);
-    const isTVEpisode = Boolean(parsedFilename && parsedFilename.show && parsedFilename.season && parsedFilename.episode);
-    if (isTVEpisode) {
-      searchRequest.name = parsedFilename.show;
-      searchRequest.reqtype = 'series';
-      const tvSeriesInfo = await omdbAPI.get(searchRequest);
-
-      if (tvSeriesInfo && tvSeriesInfo instanceof TVShow) {
-        const allEpisodes = await tvSeriesInfo.episodes();
-        const currentEpisode = _.find(allEpisodes, { season: parsedFilename.season, episode: parsedFilename.episode });
-        if (!currentEpisode) {
-          throw new IMDbIDNotFoundError();
-        }
-
-        imdbId = currentEpisode.imdbid;
-      }
-    }
-
-    if (!imdbId) {
-      searchRequest.reqtype = 'movie';
-      let searchResults;
-      try {
-        searchResults = await omdbAPI.search(searchRequest);
-      } catch (e) {
-        console.error(e);
-        return null;
-      }
-      // find the best search results utilising the Jaro-Winkler distance metric
-      const searchResultStringDistance = searchResults.results.map(result => natural.JaroWinklerDistance(searchRequest.name, result.title));
-      const bestSearchResultKey = _.indexOf(searchResultStringDistance, _.max(searchResultStringDistance));
-
-      const searchResult = searchResults.results[bestSearchResultKey] as Movie;
-      if (!searchResult) {
-        throw new IMDbIDNotFoundError();
-      }
-
-      imdbId = searchResult.imdbid;
-    }
-  }
-
-  const imdbData = await omdbAPI.get({ id: imdbId });
-  if (!imdbData) {
-    return null;
-  }
-
-  let metadata;
-  if (imdbData.type === 'movie') {
-    metadata = mapper.parseOMDbAPIMovieResponse(imdbData);
-  } else if (imdbData.type === 'series') {
-    metadata = mapper.parseOMDbAPISeriesResponse(imdbData);
-  } else if (imdbData.type === 'episode') {
-    metadata = mapper.parseOMDbAPIEpisodeResponse(imdbData);
-  } else {
-    throw new Error('Received a type we did not expect');
-  }
-
-  return metadata;
-};
 
 /**
  * Attempts a query to the OMDb API and standardizes the response
